terraform/scripts/cloud_run_ingestion/requirements.txt
flask==3.*
google-cloud-pubsub==2.*
gunicorn==22.*
